requirements.txt
phidata 
nest_asyncio 
beautifulsoup4 
openai 
duckduckgo-search 
sqlalchemy 
pgvector 
psycopg 
pypdf 
ollama
pypdf
llama-index
python-dotenv